Post controller update

+++ src/main/java/com/progparcomposant/reseausocial/controllers/InvitationController.java
@@ -1,5 +1,6 @@
 package com.progparcomposant.reseausocial.controllers;
 
+import com.progparcomposant.reseausocial.controllers.errors.ErrorMessagesEnum;
 import com.progparcomposant.reseausocial.converters.FriendshipConverter;
 import com.progparcomposant.reseausocial.converters.InvitationConverter;
 import com.progparcomposant.reseausocial.dto.FriendshipDTO;
@@ -30,6 +31,17 @@ public class InvitationController {
         this.invitationConverter = invitationConverter;
     }
 
+    @GetMapping
+    public List<InvitationDTO> findAllInvitations() {
+        Iterable<Invitation> invitations = this.invitationRepository.findAll();
+
+        if (IterableUtils.size(invitations) > 0) {
+            return this.invitationConverter.entityToDto(IterableUtils.toList(invitations));
+        } else {
+            throw new NoSuchElementException(ErrorMessagesEnum.INVITATION_NO_INVITATIONS_IN_DATABASE.getErrorMessage());
+        }
+    }
+
     @PostMapping(path = "/new")
     public InvitationDTO newInvitation(@RequestBody InvitationDTO newInvitationDTO) {
         return invitationConverter.entityToDto(this.invitationRepository.save(invitationConverter.dtoToEntity(newInvitationDTO)));
@@ -41,7 +53,7 @@ public class InvitationController {
         if (invitation.isPresent()) {
             return this.invitationConverter.entityToDto(invitation.get());
         } else {
-            throw new NoSuchElementException("Cette invitation n'existe pas");
+            throw new NoSuchElementException(ErrorMessagesEnum.INVITATION_NOT_FOUND.getErrorMessage());
         }
     }
 
@@ -51,7 +63,7 @@ public class InvitationController {
         if (IterableUtils.size(invitations) > 0) {
             return this.invitationConverter.entityToDto(IterableUtils.toList(invitations));
         } else {
-            throw new NoSuchElementException("Ce user n'a pas d'invitations");
+            throw new NoSuchElementException(ErrorMessagesEnum.INVITATION_NO_INVITATION_YET.getErrorMessage());
         }
     }
 
@@ -64,7 +76,7 @@ public class InvitationController {
             this.friendshipRepository.save(friendshipConverter.dtoToEntity(friendshipDTO));
             this.invitationRepository.deleteByFirstUserIdAndSecondUserId(invitationDTO.getFirstUserId(), invitationDTO.getSecondUserId());
         } else {
-            throw new NotFoundException("Cette invitation n'existe pas");
+            throw new NotFoundException(ErrorMessagesEnum.INVITATION_NOT_FOUND.getErrorMessage());
         }
     }
 
@@ -78,17 +90,22 @@ public class InvitationController {
             this.friendshipRepository.saveAll(this.friendshipConverter.dtoToEntity(newFriendshipDTOS));
             this.invitationRepository.deleteAllByFirstUserId(userId);
         } else {
-            throw new NoSuchElementException("Ce user n'existe pas");
+            throw new NoSuchElementException(ErrorMessagesEnum.INVITATION_NOT_FOUND.getErrorMessage());
         }
     }
 
     @DeleteMapping(path = "/delete/{firstUserId}/{secondUserId}")
     public void deleteInvitationById(@PathVariable("firstUserId") Long firstUserId, @PathVariable("secondUserId") Long secondUserId) {
-        this.invitationRepository.deleteByFirstUserIdAndSecondUserId(firstUserId, secondUserId);
+        Optional<Invitation> invitation = this.invitationRepository.findByFirstUserIdAndSecondUserId(firstUserId, secondUserId);
+        if (invitation.isPresent()) {
+            this.invitationRepository.deleteByFirstUserIdAndSecondUserId(firstUserId, secondUserId);
+        } else {
+            throw new NoSuchElementException(ErrorMessagesEnum.INVITATION_NOT_FOUND.getErrorMessage());
+        }
     }
 
     @DeleteMapping(path = "/delete/all/{userId}")
     public void deleteAllInvitationsByUserId(@PathVariable("userId") Long userId) {
-        this.invitationRepository.deleteAllByFirstUserId(userId);
+        this.invitationRepository.deleteAllByFirstUserIdOrSecondUserId(userId, userId);
     }
 }
